Move `LockableData` and `ImmutableLockableData`.
Moved to the base manipulator packages - they apply to hopper and chest
minecarts as well.
