+++ src/main/java/org/spongepowered/api/data/manipulator/mutable/LockableData.java
@@ -22,12 +22,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.api.data.manipulator.mutable.tileentity;
+package org.spongepowered.api.data.manipulator.mutable;
 
 
 import org.spongepowered.api.block.tileentity.carrier.TileEntityCarrier;
 import org.spongepowered.api.data.manipulator.DataManipulator;
-import org.spongepowered.api.data.manipulator.immutable.tileentity.ImmutableLockableData;
+import org.spongepowered.api.data.manipulator.immutable.ImmutableLockableData;
 import org.spongepowered.api.data.value.mutable.Value;
 import org.spongepowered.api.item.inventory.ItemStack;
